Redundant files in /usr/include/boinc
The situation is somewhat unfortunate in that the sample
code expects the subdirectories lib and api to host .h
files, but boinc itself does not keep up that distinction
in its destination folder with "make install".

How exactly this will develop remains unclear. The current
situation is that the files in lib and api are partially
redundant - but only partially. This may be a bug in the
BOINC install routine.